Move application form previews to flex folder
